requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
reportlab>=4.0.0
apscheduler>=3.10.0
Werkzeug>=2.3.0
python-dateutil>=2.8.0
bcrypt>=4.0.0
gunicorn>=21.0.0
pystray>=0.19.0
Pillow>=10.0.0
